+++ spec/services/share-service_spec.sh
@@ -147,7 +147,7 @@ Describe 'services'
     When call delete_service
     The status should be success
     The output should json '.version | keys == ["timestamp"]'
-    The error should include "Deleting service $service_instance"
+    The error should include "Deleting service"
     Assert not test::service_exists "$service_instance" "$source_org" "$source_space"
   End
 End