+++ spec/services/ups_with_route_service_spec.sh
@@ -108,7 +108,6 @@ Describe 'routes'
         #|  service_instance: my_route_service
         #|  hostname: $app_hostname
         #|  path: foo
-        #|  configuration: {}
       )
       put "$config"
     }